TelegramBot/requirements.txt
beautifulsoup4==4.10.0
requests==2.26.0
python-telegram-bot==13.7